refactor sidebar into thist dsfauodkg
